test/2004-e2e-move-files-to-settings-folder
