+++ e2e/selectors/Settings/Advanced/FiatOnTestnetsModal.selectors.js
