+++ e2e/selectors/Settings/Contacts/DeleteContactBottomSheet.selectors.js
@@ -1,6 +1,6 @@
-import enContent from '../../../locales/languages/en.json';
+import enContent from '../../../../locales/languages/en.json';
 
-export const DeleteContactModalSelectorsText = {
+export const DeleteContactBottomSheetSelectorsText = {
   MODAL_TITLE: enContent.address_book.delete_contact,
   DELETE_BUTTON: enContent.address_book.delete,
   CANCEL_BUTTON: enContent.address_book.cancel,
