+++ e2e/selectors/Settings/SecurityAndPrivacy/AutoLockModal.selectors.js
@@ -1,4 +1,4 @@
-import enContent from '../../../locales/languages/en.json';
+import enContent from '../../../../locales/languages/en.json';
 
 export const AutoLockModalSelectorsText = {
   AUTO_LOCK_IMMEDIATE: enContent.app_settings.autolock_immediately,
