--- e2e/selectors/Settings/SecurityAndPrivacy/ClearCookiesSection.selectors.js
@@ -1,3 +1,0 @@
-export const ClearCookiesSectionSelectorsIDs = {
-  CONTAINER: 'clear-cookies-section',
-};
