+++ e2e/specs/accounts/auto-lock.spec.js
@@ -14,7 +14,7 @@ import { loginToApp } from '../../viewHelper.js';
 import TabBarComponent from '../../pages/TabBarComponent';
 import SettingsView from '../../pages/Settings/SettingsView';
 import SecurityAndPrivacy from '../../pages/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
-import AutoLockModal from '../../pages/modals/AutoLockModal';
+import AutoLockModal from '../../pages/Settings/SecurityAndPrivacy/AutoLockModal';
 import Assertions from '../../utils/Assertions.js';
 import WalletView from '../../pages/wallet/WalletView.js';
 import LoginView from '../../pages/LoginView.js';
