+++ e2e/specs/settings/fiat-on-testnets.spec.js
@@ -10,7 +10,7 @@ import NetworkListModal from '../../pages/Network/NetworkListModal';
 import WalletView from '../../pages/wallet/WalletView';
 import NetworkEducationModal from '../../pages/Network/NetworkEducationModal';
 import AdvancedSettingsView from '../../pages/Settings/AdvancedView';
-import FiatOnTestnetsModal from '../../pages/modals/FiatOnTestnetsModal.js';
+import FiatOnTestnetsModal from '../../pages/Settings/Advanced/FiatOnTestnetsModal.js';
 import Assertions from '../../utils/Assertions.js';
 import TestHelpers from '../../helpers.js';
